@@ -1,67 +1,106 @@
 name: shop_order
 description: A new Flutter project.
 
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+# In Windows, build-name is used as the major, minor, and patch parts
+# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: '>=2.18.4 <3.0.0'
 
-
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
 
-  # Base
-  cupertino_icons: ^1.0.3
-  nb_utils: ^4.6.23
-  intl: ^0.17.0
-  flutter_vector_icons: ^2.0.0
-  google_fonts: ^3.0.1
 
-  # STATE MANAGEMENT
-  mobx: ^2.0.3
-  flutter_mobx: ^2.0.1
-
-  #ShopHop
-  flutter_svg: ^1.1.6
-  carousel_slider: ^4.0.0
-  flutter_slidable: ^0.6.0
-  url_launcher: ^6.0.3
-  percent_indicator: ^4.2.2
-  cached_network_image: ^3.1.0
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^1.0.2
   device_preview: ^1.1.0
-
+  nb_utils: ^4.6.23
+  flutter_rating_bar: ^4.0.1
+  image_picker: ^0.8.6
+  country_code_picker: ^2.0.2
+  cached_network_image: ^3.2.2
+  flutter_mobx: ^2.0.6+4
+  google_fonts: ^3.0.1
+  flutter_slidable: ^2.0.0
+  timeline_tile: ^2.0.0
+  dots_indicator: ^2.1.0
+  font_awesome_flutter: ^10.2.1
+  flutter_format_money_vietnam: ^0.0.2
+  shared_preferences: ^2.0.15
+  fluttertoast: ^8.1.1
+  
+  # cloud_firestore: ^4.0.4
+  # onesignal_flutter: ^3.4.2
+  # pin_code_fields: ^7.4.0
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mobx_codegen: ^2.0.2
-  build_runner: ^2.1.0
 
-  flutter_launcher_icons: ^0.10.0
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: ^2.0.0
 
-flutter_icons:
-  android: true
-  ios: true
-  image_path: "images/shophop/ic_app_icon.png"
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
 
+# The following section is specific to Flutter packages.
 flutter:
 
-
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
   uses-material-design: true
-
   assets:
-   - images/shophop/
-   - images/
-   - images/shophop/img/
-   - assets/shophop_data/
-   - images/shophop/img/products/
-   - images/shophop/img/products/shoes/
-   - images/shophop/img/products/dress/
-   - images/shophop/img/products/banners/
-   - images/shophop/img/products/decor/
+    - assets/
+    - assets/products/
+    - assets/walk_through/
+    - assets/category/
+    - assets/slider/
+    - assets/app/
 
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware
+
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
   # fonts:
   #   - family: Schyler
   #     fonts:
